I've added some validations

+++ src/sockets/sockets.controller.ts
@@ -13,34 +13,51 @@ class SocketConnection {
       console.log('A user connected');
 
       // get all the current todos
-      const allTodos = await Todo.find({});
+      const allTodos = await Todo.find({}).sort({
+        createdAt: 'desc',
+      });
+
       this.#io.emit('getTodos', allTodos);
 
       // create a new todo in the database
       socket.on('createTodo', async (payload) => {
+        if (!payload.title) {
+          return;
+        }
         await Todo.create(payload);
 
-        const allTodos = await Todo.find({});
+        const allTodos = await Todo.find({}).sort({
+          createdAt: 'desc',
+        });
         this.#io.emit('getTodos', allTodos);
       });
 
       // delete a todo from the database
       socket.on('deleteTodo', async (payload) => {
+        if (!payload.todoId) {
+          return;
+        }
+
         const { todoId } = payload;
         await Todo.findByIdAndDelete(todoId);
 
-        const allTodos = await Todo.find({});
+        const allTodos = await Todo.find({}).sort({
+          createdAt: 'desc',
+        });
         this.#io.emit('getTodos', allTodos);
       });
 
       socket.on('updateTodo', async (payload) => {
-        console.log(payload);
-
+        if (!payload.id) {
+          return;
+        }
         await Todo.findByIdAndUpdate(payload.id, {
           title: payload.title,
         });
 
-        const allTodos = await Todo.find({});
+        const allTodos = await Todo.find({}).sort({
+          createdAt: 'desc',
+        });
         this.#io.emit('getTodos', allTodos);
       });
     });
